refactor(signals): route emit_signal through backend facade
Move emit_signal from products/signals/backend/api.py to
products/signals/backend/facade/api.py so cross-boundary callers
(endpoints product, core temporal workers, scout harness) go through a
real facade per product-isolation guidelines.

The temporal workflow imports emit_signal needs are deferred into the
function body. At module scope they formed a latent circular import
(api -> temporal package -> reingestion -> api) that only resolved by
import-order luck and dragged the temporal workflow stack onto the
Django startup path. The facade now imports nothing temporal at module
scope, so importing it from the API/URL hot path is cheap and safe.

Pure rename + deferred imports; no behavior change.
