+++ products/endpoints/backend/tests/test_endpoint_execution.py
@@ -2186,7 +2186,7 @@ class TestEndpointExecution(ClickhouseTestMixin, APIBaseTest):
         endpoint = self._make_simple_hogql_endpoint("failure_signal_swallow")
 
         with mock.patch(
-            "products.signals.backend.api.emit_signal",
+            "products.signals.backend.facade.api.emit_signal",
             side_effect=RuntimeError("signal layer exploded"),
         ):
             _emit_endpoint_failure_signal(self.team, endpoint, RuntimeError("original"), materialized=False, version=1)
